+++ src/Anonymization/Anonymizer/Context.php
@@ -7,11 +7,14 @@ namespace MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer;
 class Context
 {
     public readonly string $salt;
+    public readonly string $basePath;
 
     public function __construct(
         ?string $salt = null,
+        ?string $basePath = null,
     ) {
         $this->salt = $salt ?? self::generateRandomSalt();
+        $this->basePath = $basePath ?? \getcwd();
     }
 
     public static function generateRandomSalt(): string